convert some more floats to ints

+++ core/src/main/java/org/elasticsearch/cluster/routing/allocation/allocator/BalancedShardsAllocator.java
@@ -871,7 +871,7 @@ public class BalancedShardsAllocator extends AbstractComponent implements Shards
             return routingNode;
         }
 
-        public float totalShardWeight() {
+        public int totalShardWeight() {
             return totalShardWeight;
         }
